adição de criar projeto e atualização de nome da classe
